requirements/base.txt
pytz==2020.4  # https://github.com/stub42/pytz
python-slugify==4.0.1  # https://github.com/un33k/python-slugify
Pillow==8.0.1  # https://github.com/python-pillow/Pillow
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.2.0  # https://github.com/evansd/whitenoise
redis==3.5.3  # https://github.com/andymccurdy/redis-py
hiredis==1.1.0  # https://github.com/redis/hiredis-py
celery==4.4.6  # pyup: < 5.0,!=4.4.7  # https://github.com/celery/celery
django-celery-beat==2.1.0  # https://github.com/celery/django-celery-beat
flower==0.9.5  # https://github.com/mher/flower
uvicorn[standard]==0.13.4  # https://github.com/encode/uvicorn

# Django
# ------------------------------------------------------------------------------
django==3.1.3  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-crispy-forms==1.10.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.12.1  # https://github.com/jazzband/django-redis
django-timezone-field==4.0  # https://github.com/mfogel/django-timezone-field/
django-cors-headers==3.5.0 # https://github.com/adamchainz/django-cors-headers
django-csp==3.7  # https://github.com/mozilla/django-csp
# Django REST Framework
djangorestframework==3.12.2  # https://github.com/encode/django-rest-framework
drf-yasg==1.20.0 # https://github.com/axnsan12/drf-yasg
django-debug-toolbar~=3.1.1

mock~=4.0.2
pytest~=6.1.2
Faker~=8.8.1
factory-boy
